extract TemplateSearchPicker to fix workflow template pagination bug
